requirements.txt
requests==2.25.1
beautifulsoup4==4.10.0
Pillow==8.3.2
PyPDF2==1.26.0
